+++ js/application.js
@@ -1,5 +1,6 @@
 import welcomeScreen from './welcome/welcome-screen';
 import newGameScreen from './game/game-screen';
+import resultScreen from './result/result-screen';
 
 export default class Application {
 
@@ -10,5 +11,8 @@ export default class Application {
   static showGame() {
     newGameScreen.init();
   }
-}
 
+  static showStats() {
+    resultScreen.init();
+  }
+}
