+++ js/data/data.js
@@ -81,7 +81,7 @@ const musicData = new Set([
     artist: `Gunnar Olsen`,
     name: `Home Stretch`,
     image: `https://f4.bcbits.com/img/0004181452_10.jpg`,
-    src: `https://www.youtube.com/audiolibrary_download?vid=bcbe5be936a32fb1`,
+    src: `https://www.youtube.com/audiolibrary_download?vid=6feb0654949ef64c`,
     genre: `Electronic`
   }
 ]);
