+++ js/data/genre-screen-data.js
@@ -24,7 +24,7 @@ const generateGenreQuestion = (numberOfAnswers) => {
   const questionData = [...[...musicData].keys()].filter((it) => {
     return !initialState.genreQuestions.has(it);
   });
-  const questionNumber = questionData[getRandomInt(0, questionData.length - 1)];
+  const questionNumber = questionData[getRandomInt(0, questionData.length)];
 
   const generateAnswers = () => {
     const answers = [];
